Fixing LOAD DATA LOCAL INFILE for SSCursor* types
There isn't a compelling reason for using an SSCursor for such a command
(as it is really intented for queries that return large result sets),
but I use some wrapper code that always uses these types of cursors
(since most interactions are heavy SELECTs), so prefer it to work
in all cases. Added a test to demonstrate.