Improve documentation with errors and links
I added several links to the PEP 249 documentation. I also documented
the various exceptions that can be raised and, clarified the meaning
of some parameters.

Overall, readers will be able to cross-reference methods with the
specification, and will see implementation-specific info. This will
make it clear that this library implements the full specification,
which is important because there is almost no direct documentation
of this library.