Add interface for non-file LOAD DATA LOCAL INFILE (#376)
LOAD DATA LOCAL INFILE is a rapid load mechanism in MySQL. ETL,
machine learning, and specialized accelerators are some of the
reasons that a rapid generation of data can occur. These are
application domains where Python is frequently used. Placing
generated data into to a file before a load is an additional
complication. This commit provides an alternate to  temporary
file generation.

In this commit the load_infile argument is expanded from a
boolean into a dictionary. In true pythonic style, a dictionary
implements `__getitem__` and during the client side LOAD DATA
LOCAL INFILE implementation it is the filename that is looked
up in the dictionary.

The object retrieved from the dictionary will have have its
`read` method called (argument max_packet_size) to retrieve data
as if it was the file being uploaded. The `read` method will
return CSV, or otherwise separated, data as bytes. When the
`read` method is complete it will return None.

closes #376